requirements-auth.txt
# Production Requirements for RAG System
# Core Flask and web server components
flask==3.0.0
gunicorn==21.2.0
gevent==23.9.1

# Authentication and security
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0
PyJWT==2.8.0
flask-limiter==3.5.0

# Environment and configuration
python-dotenv==1.0.0

# Enhanced requirements from base
flask-cors==4.0.0
requests==2.31.0
chromadb==0.4.18
sentence-transformers==2.2.2
langchain==0.0.340
langchain-community==0.0.2
openai==1.3.7
google-generativeai==0.3.2
rank-bm25==0.2.2

# Production optimizations
redis==5.0.1
psycopg2-binary==2.9.9
sqlalchemy==2.0.23

# Monitoring and logging
sentry-sdk[flask]==1.38.0
prometheus-flask-exporter==0.22.4

# File processing
python-magic==0.4.27
PyPDF2==3.0.1
python-docx==1.1.0
openpyxl==3.1.2

# ML and AI libraries (production versions)
torch==2.1.1
transformers==4.35.2
numpy==1.24.4
scikit-learn==1.3.2

# Utilities
tqdm==4.66.1
colorama==0.4.6